% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.R
\name{get_bias}
\alias{get_bias}
\alias{get_rmse}
\alias{get_corr}
\title{Bias, RMSE and correlation}
\usage{
get_bias(est, true)

get_rmse(est, true)

get_corr(est, true)
}
\arguments{
\item{est, true}{Estimates and true values.}
}
\description{
Bias, RMSE and correlation
}
