% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aliases.R
\name{julia_aliases}
\alias{julia_aliases}
\alias{setCond}
\alias{InputData}
\alias{InputData4R}
\alias{InputPara}
\alias{GibbsMlIrt}
\alias{GibbsRtIrtNull}
\alias{GibbsRtIrt}
\alias{GibbsRtIrtLatent}
\alias{GibbsRtIrtLatentQr}
\alias{GibbsRtIrtCross}
\alias{GibbsRtIrtCrossQr}
\alias{sampleGibbs}
\alias{getDic}
\alias{checkConvergence}
\alias{comparePara}
\alias{getBias}
\alias{getRmse}
\alias{getCorr}
\alias{setTrueParaMlIrt}
\alias{setTrueParaRtIrt}
\alias{setTrueParaRtIrtLatent}
\alias{setTrueParaRtIrtCross}
\alias{setDataMlIrt}
\alias{setDataRtIrt}
\alias{setDataRtIrtNull}
\alias{setDataRtIrtLatent}
\alias{setDataRtIrtCross}
\alias{runSimulation}
\title{Julia-compatible aliases}
\usage{
setCond(
  nSubj = 2000,
  nItem = 15,
  nFeat = 3,
  nIter = 5000,
  nChain = 4,
  nBurnin = NULL,
  nThin = 1,
  nRep = 10,
  qRa = 0.5,
  qRt = 0.5
)

InputData(Y, T = NULL, X = NULL)

InputData4R(Y, kappa = Y - 0.5, logT = NULL, X = NULL)

InputPara(...)

GibbsMlIrt(Cond, Data, truePara = NULL, Para = NULL)

GibbsRtIrtNull(Cond, Data, truePara = NULL, Para = NULL)

GibbsRtIrt(Cond, Data, truePara = NULL, Para = NULL)

GibbsRtIrtLatent(Cond, Data, truePara = NULL, Para = NULL)

GibbsRtIrtLatentQr(Cond, Data, truePara = NULL, Para = NULL)

GibbsRtIrtCross(Cond, Data, truePara = NULL, Para = NULL)

GibbsRtIrtCrossQr(Cond, Data, truePara = NULL, Para = NULL)

sampleGibbs(MCMC, cov2one = NULL, ...)

getDic(MCMC)

checkConvergence(MCMC, ...)

comparePara(MCMC, par = "a")

getBias(est, true)

getRmse(est, true)

getCorr(est, true)

setTrueParaMlIrt(Cond)

setTrueParaRtIrt(Cond, trueCorr = 0.3)

setTrueParaRtIrtLatent(Cond)

setTrueParaRtIrtCross(Cond)

setDataMlIrt(Cond, truePara)

setDataRtIrt(Cond, truePara)

setDataRtIrtNull(Cond, truePara)

setDataRtIrtLatent(Cond, truePara, type = "norm")

setDataRtIrtCross(Cond, truePara, type = "norm")

runSimulation(
  Cond,
  truePara,
  funcData = setDataRtIrt,
  funcGibbs = GibbsRtIrt,
  par = c("a", "b", "theta"),
  ...
)
}
\arguments{
\item{nSubj, nItem, nFeat, nIter, nChain, nBurnin, nThin, nRep, qRa, qRt}{As in \code{setCond()}.}

\item{Y, T, X, kappa, logT}{Data, as in \code{InputData()} / \code{InputData4R()}.}

\item{Cond, Data, truePara, Para, MCMC, par, est, true, ...}{As in the Julia package.}

\item{cov2one}{Julia's setting: \code{TRUE} fixes the variance of speed at 1 (\code{speed_var = "fixed"}), \code{FALSE} frees it.}

\item{trueCorr, type, funcData, funcGibbs}{As in the Julia functions.}
}
\description{
The function names of ExtendedRtIrtModeling.jl, mapped to the snake_case functions:
}
\details{
\tabular{ll}{
   Julia \tab birtRcpp \cr
   \code{setCond(nSubj, nItem, nFeat, nIter, nChain, nBurnin, nThin, nRep, qRa, qRt)} \tab \code{\link[=set_cond]{set_cond()}} \cr
   \code{InputData(Y, T, X)}, \code{InputData4R(Y, kappa, logT, X)} \tab \code{\link[=input_data]{input_data()}} (\code{resp}, \code{time}, \code{cov}, \code{log_time}) \cr
   \code{InputPara(...)} \tab \code{\link[=input_para]{input_para()}} \cr
   \code{GibbsMlIrt}, \code{GibbsRtIrtNull}, \code{GibbsRtIrt} \tab \code{\link[=mlirt]{mlirt()}}, \code{\link[=rtirt_null]{rtirt_null()}}, \code{\link[=rtirt_latreg]{rtirt_latreg()}} \cr
   \code{GibbsRtIrtLatent}, \code{GibbsRtIrtLatentQr} \tab \code{\link[=rtirt_latent]{rtirt_latent()}}, \code{quantile = NULL} / \code{Cond$q_rt} \cr
   \code{GibbsRtIrtCross}, \code{GibbsRtIrtCrossQr} \tab \code{\link[=rtirt_cross]{rtirt_cross()}}, \code{quantile = NULL} / \code{Cond$q_rt} \cr
   \verb{sample!(MCMC)} \tab \code{sampleGibbs(MCMC)} = \code{\link[=gibbs]{gibbs()}} \cr
   \code{coef(MCMC)}, \code{precis(MCMC)} \tab \link[=summary.rtirt]{summary()}, \code{\link[=precis]{precis()}} \cr
   \code{getDic}, \code{checkConvergence}, \code{comparePara} \tab \code{\link[=dic]{dic()}}, \code{\link[=convergence]{convergence()}}, \code{\link[=compare_para]{compare_para()}} \cr
   \code{getBias}, \code{getRmse}, \code{getCorr} \tab \code{\link[=get_bias]{get_bias()}}, \code{\link[=get_rmse]{get_rmse()}}, \code{\link[=get_corr]{get_corr()}} \cr
   \verb{setTruePara<Model>}, \verb{setData<Model>} \tab \code{\link[=sim_para]{sim_para()}}, \code{\link[=sim_data]{sim_data()}} \cr
   \verb{runSimulation(Cond, truePara; funcData, funcGibbs)} \tab \code{\link[=run_simulation]{run_simulation()}} \cr
}


The constructors take \verb{(Cond, Data, truePara, Para)} as in Julia and keep the Julia
defaults: the speed variance is fixed at 1 (\code{speed_var = "fixed"}) in \code{GibbsRtIrtNull},
\code{GibbsRtIrt} and \code{GibbsRtIrtCross} (\code{sampleGibbs(M, cov2one = FALSE)} frees it), and the
\code{Qr} versions use the quantile level \code{Cond$q_rt}. \code{setCond()} keeps the Julia default
dimensions (2000 persons, 15 items, 3 covariates), which must match the data.
}
