% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{reliability}
\alias{reliability}
\alias{reliability.rtirt}
\title{Empirical reliability}
\usage{
reliability(object, ...)

\method{reliability}{rtirt}(object, ...)
}
\arguments{
\item{object}{A sampled model.}

\item{...}{Unused.}
}
\description{
var(EAP) / (var(EAP) + mean PSD^2) of ability and speed.
}
