% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{scores}
\alias{scores}
\alias{scores.rtirt}
\title{Person scores}
\usage{
scores(object, ...)

\method{scores}{rtirt}(object, ...)
}
\arguments{
\item{object}{A sampled model.}

\item{...}{Unused.}
}
\value{
A data frame with \code{id}, \code{ability}, \code{ability_psd} and (RT models) \code{speed}, \code{speed_psd}.
}
\description{
Posterior means (EAP) and posterior SDs of ability and speed, one row per person.
}
